mito-ai: add azure capabilities test
